+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # time-checker changelog
 
+## [0.1.1](https://github.com/ivanoff/time-checker/tree/0.1.1) (2017-01-..)
+[Full Changelog](https://github.com/ivanoff/time-checker/compare/0.0.1...0.1.1)
+
+**What Was Done:**
+
+- add options parameters
+- add usage example
+
+
 ## [0.0.1](https://github.com/ivanoff/time-checker/tree/0.0.1) (2017-01-11)
 [Full Changelog](https://github.com/ivanoff/time-checker/compare/0.0.1...0.0.1)